Exposes codec
