+++ package.json
@@ -2,7 +2,7 @@
   "author": "Transcend Inc.",
   "name": "@transcend-io/cli",
   "description": "Small package containing useful typescript utilities.",
-  "version": "4.69.0",
+  "version": "4.70.0",
   "homepage": "https://github.com/transcend-io/cli",
   "repository": {
     "type": "git",
